requirements.txt
click
rich
Jinja2
jinja2-strcase
GitPython